requirements.txt
absl-py==0.6.1
astor==0.7.1
certifi==2018.11.29
chardet==3.0.4
gast==0.2.1.post0
grpcio==1.17.1
h5py==2.9.0
idna==2.8
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
Markdown==3.0.1
mne==0.17.0
mock==2.0.0
numpy==1.15.4
OpenBCI-Python==1.0.0
pbr==5.1.1
Pillow==5.4.1
protobuf==3.6.1
pygame==1.9.4
pylsl==1.13.1
pyserial==3.4
python-osc==1.7.0
PyYAML==3.13
requests==2.21.0
scikit-learn==0.20.2
scipy==1.2.0
six==1.12.0
sklearn==0.0
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
torch==1.0.0
torchvision==0.2.1
urllib3==1.24.1
Werkzeug==0.14.1
xmltodict==0.11.0
Yapsy==1.12.0
